+++ ts/features/itwallet/discovery/components/ItwDiscoveryInfoComponent.tsx
@@ -8,7 +8,7 @@ import { tosConfigSelector } from "../../../tos/store/selectors/index.ts";
 import { trackItwIntroBack, trackOpenItwTos } from "../../analytics/index.ts";
 import { itwIsActivationDisabledSelector } from "../../common/store/selectors/remoteConfig.ts";
 import { selectIsLoading } from "../../machine/eid/selectors.ts";
-import { ItwEidIssuanceMachineContext } from "../../machine/provider.tsx";
+import { ItwEidIssuanceMachineContext } from "../../machine/eid/provider.tsx";
 import { generateLinkRuleWithCallback } from "../../common/utils/markdown.tsx";
 import { IOScrollView } from "../../../../components/ui/IOScrollView.tsx";
 import { useHeaderSecondLevel } from "../../../../hooks/useHeaderSecondLevel.tsx";