+++ ts/features/itwallet/identification/screens/cie/ItwCieCardReaderL2Screen.tsx
@@ -46,10 +46,10 @@ import {
   selectAuthUrlOption,
   selectCiePin
 } from "../../../machine/eid/selectors";
-import { ItwEidIssuanceMachineContext } from "../../../machine/provider";
 import { ItwParamsList } from "../../../navigation/ItwParamsList";
 import { ITW_ROUTES } from "../../../navigation/routes";
 import * as Cie from "../../components/cie";
+import { ItwEidIssuanceMachineContext } from "../../../machine/eid/provider";
 
 // the timeout we sleep until move to consent form screen when authentication goes well
 const WAIT_TIMEOUT_NAVIGATION = 1700 as Millisecond;
@@ -154,7 +154,7 @@ const LoadingSpinner = (
   <LoadingScreenContent contentTitle={I18n.t("global.genericWaiting")} />
 );
 
-export const ItwCieCardReaderScreen = () => {
+export const ItwCieCardReaderL2Screen = () => {
   const navigation = useNavigation<StackNavigationProp<ItwParamsList>>();
   const env = useIOSelector(selectItwEnv);
   const { ISSUANCE_REDIRECT_URI } = getEnv(env);