+++ ts/features/itwallet/machine/eid/actors.ts
@@ -1,6 +1,7 @@
 import cieManager from "@pagopa/react-native-cie";
 import * as O from "fp-ts/lib/Option";
 import { fromPromise } from "xstate";
+import { Trust } from "@pagopa/io-react-native-wallet-v2";
 import { useIOStore } from "../../../../store/hooks";
 import { sessionTokenSelector } from "../../../authentication/common/store/selectors";
 import { assert } from "../../../../utils/assert";
@@ -48,6 +49,10 @@ export type GetWalletAttestationActorParams = {
   isL3IssuanceEnabled?: boolean;
 };
 
+export type VerifyTrustFederationParams = {
+  isL3IssuanceEnabled?: boolean;
+};
+
 /**
  * Creates the actors for the eid issuance machine
  * @param env - The environment to use for the IT Wallet API calls
@@ -58,6 +63,40 @@ export const createEidIssuanceActorsImplementation = (
   env: Env,
   store: ReturnType<typeof useIOStore>
 ) => ({
+  verifyTrustFederation: fromPromise<void, VerifyTrustFederationParams>(
+    async ({ input }) => {
+      // If the L3 issuance is not enabled, we don't need to verify the trust federation
+      if (!input.isL3IssuanceEnabled) {
+        return;
+      }
+
+      // Evaluate the issuer trust
+      const trustAnchorEntityConfig =
+        await Trust.Build.getTrustAnchorEntityConfiguration(
+          env.WALLET_TA_BASE_URL
+        );
+      const trustAnchorKey = trustAnchorEntityConfig.payload.jwks.keys[0];
+
+      // Create the trust chain for the PID provider
+      // TODO: [SIW-2530] Move "1-0" to WALLET_PID_PROVIDER_BASE_URL after migrating to the new API
+      const builtChainJwts = await Trust.Build.buildTrustChain(
+        new URL("1-0", env.WALLET_PID_PROVIDER_BASE_URL).toString(),
+        trustAnchorKey
+      );
+
+      // Perform full validation on the built chain
+      await Trust.Verify.verifyTrustChain(
+        trustAnchorEntityConfig,
+        builtChainJwts,
+        {
+          connectTimeout: 10000,
+          readTimeout: 10000,
+          requireCrl: true
+        }
+      );
+    }
+  ),
+
   createWalletInstance: fromPromise<string>(async () => {
     const sessionToken = sessionTokenSelector(store.getState());
     assert(sessionToken, "sessionToken is undefined");