+++ ts/features/pn/navigation/navigator.tsx
@@ -3,6 +3,8 @@ import { MessageAttachmentScreen } from "../screens/MessageAttachmentScreen";
 import { MessageDetailsScreen } from "../screens/MessageDetailsScreen";
 import { PaidPaymentScreen } from "../screens/PaidPaymentScreen";
 import { PNActivationBannerFlowScreen } from "../reminderBanner/screens/PnReminderBannerFlow";
+import { SendEngagementScreen } from "../aar/screen/SendEngagementScreen";
+import { SendQrScanFlow } from "../qrCodeScan/screens/QrScanFlow";
 import { PnParamsList } from "./params";
 import PN_ROUTES from "./routes";
 
@@ -29,5 +31,22 @@ export const PnStackNavigator = () => (
         headerShown: false
       }}
     />
+    <Stack.Screen
+      name={PN_ROUTES.QR_SCAN_FLOW}
+      component={SendQrScanFlow}
+      options={{
+        headerShown: false
+      }}
+    />
+    <Stack.Group
+      screenOptions={{
+        presentation: "modal"
+      }}
+    >
+      <Stack.Screen
+        name={PN_ROUTES.ENGAGEMENT_SCREEN}
+        component={SendEngagementScreen}
+      />
+    </Stack.Group>
   </Stack.Navigator>
 );